+++ tests/test_communication.py
@@ -1,13 +1,26 @@
 from os import environ
+import socket
+
 import pytest
-from serial import Serial, SerialException, SerialTimeoutException
+from serial import Serial
 from geocompy.communication import (
     get_dummy_logger,
     open_serial,
+    open_socket,
     SerialConnection,
+    SocketConnection,
     crc16_bitwise,
     crc16_bytewise
 )
+
+
+@pytest.fixture
+def sock() -> socket.socket:
+    return socket.socket(
+        socket.AF_INET,
+        socket.SOCK_STREAM,
+        socket.IPPROTO_TCP
+    )
 
 
 portname = environ.get("GEOCOMPY_TEST_PORT_CLIENT", "")
@@ -24,12 +37,93 @@ if faultyportname == "":  # pragma: no coverage
         "'GEOCOMPY_TEST_PORT_FAULTY' environment variable"
     )
 
+tcpport = environ.get("GEOCOMPY_TEST_TCPPORT_SERVER", "")
+if faultyportname == "":  # pragma: no coverage
+    raise ValueError(
+        "Echo server tcp port must be set in "
+        "'GEOCOMPY_TEST_TCPPORT_SERVER' environment variable"
+    )
+
 
 class TestDummyLogger:
     def test_get_dummy_logger(self) -> None:
         log = get_dummy_logger()
         assert log.name == "geocompy.dummy"
         assert len(log.handlers) == 1
+
+
+class TestSocketConnection:
+    def test_init(self, sock: socket.socket) -> None:
+        sock.connect(("127.0.0.1", int(tcpport)))
+        with SocketConnection(sock) as client:
+            assert client.is_open()
+
+        client.close()
+        assert not client.is_open()
+
+    def test_open_socket(self) -> None:
+        with open_socket("127.0.0.1", int(tcpport), "tcp") as soc:
+            assert soc.is_open()
+
+        with pytest.raises(Exception):
+            open_socket("127.0.0.1", int(tcpport), "rfcomm", timeout=1)
+
+        with pytest.raises(ValueError):
+            open_socket(
+                "127.0.0.1",
+                int(tcpport),
+                "mistake",  # type: ignore[arg-type]
+                timeout=1
+            )
+
+    def test_messaging(self) -> None:
+        with open_socket(
+            "127.0.0.1",
+            int(tcpport),
+            "tcp"
+        ) as soc:
+            soc.is_open()
+            request = "Test"
+            assert soc.exchange(request) == request
+
+            soc.send("ascii")
+            assert soc.receive() == "ascii"
+
+            assert soc.exchange_binary(b"00\r\n") == b"00"
+
+            soc.reset()
+
+        with pytest.raises(ConnectionError):
+            soc.send("closed")
+
+        with pytest.raises(ConnectionError):
+            soc.receive()
+
+        with open_socket(
+            "127.0.0.1",
+            int(tcpport),
+            "tcp",
+            sync_after_timeout=True
+        ) as soc:
+            soc.send("msg1")
+            soc.send("msg2")
+            soc.send("msg3")
+            soc._timeout_counter = 3
+            assert soc.exchange("recovered") == "recovered"
+
+            with soc.timeout_override(1):
+                with pytest.raises(TimeoutError):
+                    soc.receive()
+
+                assert soc._timeout_counter == 1
+
+                with pytest.raises(TimeoutError):
+                    soc.receive()
+
+                assert soc._timeout_counter == 2
+
+        with pytest.raises(ConnectionError):
+            soc.receive()
 
 
 class TestSerialConnection:
@@ -62,10 +156,10 @@ class TestSerialConnection:
 
             com.reset()
 
-        with pytest.raises(SerialException):
+        with pytest.raises(ConnectionError):
             com.send("closed")
 
-        with pytest.raises(SerialException):
+        with pytest.raises(ConnectionError):
             com.receive()
 
         with open_serial(
@@ -79,12 +173,12 @@ class TestSerialConnection:
             assert com.exchange("recovered") == "recovered"
 
             with com.timeout_override(1):
-                with pytest.raises(SerialTimeoutException):
+                with pytest.raises(TimeoutError):
                     com.receive()
 
                 assert com._timeout_counter == 1
 
-                with pytest.raises(SerialTimeoutException):
+                with pytest.raises(TimeoutError):
                     com.receive()
 
                 assert com._timeout_counter == 2